Fix crash on system lock/unlock
Updated pulled from SharedLibraryNG - specifically from the mRemote3G
project where a lot of debugging was performed to avoid this crash in
the main mRemoteNG/mRemote3G app (without VncSharp connections being
made). However this didn't resolve the problems 100% of the time it
seems.

So far testing with this code in the VncSharp lib has been nothing but
successful though.